move tests to separate file, remove the now obsolete collect tests (replaced with parser tests)
